- BaseEntity 수정

+++ src/main/java/com/example/demo/service/BaseService.java
@@ -10,7 +10,7 @@ import java.util.List;
 // R : 이 추상 클래스를 상속 받은 자식 클래스에서 사용할 리포지토리
 // 사용 방법 : public class 클래스명 extends MainService<엔티티 타입, 리포지토리명>
 @RequiredArgsConstructor
-public abstract class BaseService<T, R extends JpaRepository<T, Integer>> {
+public abstract class BaseService<T, R extends JpaRepository<T, Long>> {
 
     // 이 추상 클래스로 인해 기본적인 CRUD는 각 서비스 클래스에 작성하지 않아도 됨.
 
@@ -18,7 +18,7 @@ public abstract class BaseService<T, R extends JpaRepository<T, Integer>> {
     private final R repository;
 
     // 단일 조회
-    public T get(Integer id) {
+    public T get(Long id) {
         return repository.findById(id).get();
     }
 
@@ -38,7 +38,7 @@ public abstract class BaseService<T, R extends JpaRepository<T, Integer>> {
     }
 
     // 삭제
-    public void delete(Integer id) {
+    public void delete(Long id) {
         repository.deleteById(id);
     }
 }
